user-service/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
flask-cors==4.0.1